refactor(util): make EncodeCoinbaseHeightPush uint32; add push-boundary tests
Address review feedback on #1338.

ChiR1: EncodeCoinbaseHeightPush/encodeScriptNumMinimal took int64 but only
have push_int64 parity for non-negative values, and encodeScriptNumMinimal
panicked on math.MinInt64 (negation overflow left an empty slice that the
trailing index read dereferenced). Consensus callers were already safe
(guarded to [0, MaxUint32] and uint32), but the exported int64 signature was
a trap for future callers. Change both to uint32, which makes the
non-negative contract explicit in the type and deletes the dead negative
branches and the panic path.

ChiR2: add reject cases for truncated pushes and OP_PUSHDATA1/2/4 headers
(03bfea, 4c, 4c03bfea, 4d03, 4e030000, 4e03000000) so the parser's
no-panic/no-overflow bounds checks on attacker-controlled scriptSig data are
locked down by tests, not just by inspection.

+++ model/GetCoinbaseParts.go
@@ -182,7 +182,7 @@ func makeCoinbase1(height uint32, coinbaseText string) []byte {
 	// BIP34 block height, encoded canonically (CScript() << nHeight) so it is minimally encoded and
 	// accepted by SV Node and by Teranode's own coinbase-height parser. A fixed 3-byte push is
 	// non-minimal below height 32768 and truncates above 16777215.
-	arbitraryData = append(arbitraryData, util.EncodeCoinbaseHeightPush(int64(height))...)
+	arbitraryData = append(arbitraryData, util.EncodeCoinbaseHeightPush(height)...)
 	arbitraryData = append(arbitraryData, []byte(coinbaseText)...)
 
 	// Arbitrary data should leave enough space for the extra nonce
